build finally works on MSVC

+++ base/plugins/iscore-plugin-scenario/Scenario/Document/Constraint/ViewModels/FullView/FullViewConstraintHeader.cpp
@@ -31,7 +31,7 @@ AddressBarItem *FullViewConstraintHeader::bar() const
 
 QRectF FullViewConstraintHeader::boundingRect() const
 {
-    return {0, 0, m_width, ConstraintHeader::headerHeight()};
+    return {0, 0, m_width, qreal(ConstraintHeader::headerHeight())};
 }
