+++ base/plugins/iscore-plugin-scenario/Scenario/Document/DisplayedElements/DisplayedElementsPresenter.cpp
@@ -157,7 +157,7 @@ void DisplayedElementsPresenter::on_displayedConstraintHeightChanged(double size
 {
     m_model->updateRect(
     {
-        ScenarioLeftSpace,
+        qreal(ScenarioLeftSpace),
         0,
         m_constraintPresenter->abstractConstraintViewModel().model().duration.defaultDuration().toPixels(m_constraintPresenter->model().fullView()->zoom()),
         size
